Gathered both random vectors plot in one
